+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Support for HEP-1 Value Annotations.
 - Added this change log.
 
 [Unreleased]: https://github.com/aperezdc/hipack-js/compare/v0.1.2...HEAD